policelawhelper/requirements.txt
django
djangorestframework
Pygments
Markdown
djongo
gunicorn
jieba
pika
mysqlclient
lxml
djangorestframework
Django
django-cors-headers
django-filter
django-crispy-forms
django-cors-headers
django-picklefield
django-mptt
django-import-export
django-advanced-filters
django-bulk-update
urllib3
pillow
requests